fix(clip): correct match_cakes return type for base_price
The migration created the table and function without error, but the function
failed at call time:

    Returned type integer does not match expected type numeric in column 6

Column 6 is base_price. `RETURNS TABLE` declared it NUMERIC while
products.base_price is INTEGER, and Postgres requires an exact match rather than
performing an implicit widening. This was my error: the type was assumed instead
of checked.

Fixes:
- add_clip_image_search.sql: base_price declared INTEGER, and the select casts
  p.base_price::INTEGER so the guarantee is explicit in the query too
- adds fix_match_cakes_returntype.sql, a standalone CREATE OR REPLACE that
  repairs an already-deployed function without touching the table or its data

The remaining five return columns were then checked against the live schema via
the PostgREST OpenAPI document rather than by eye: id/uuid, name and category/
character varying, product_type/text all match their declared TEXT, and
base_price/integer now matches INTEGER.

+++ backend/migrations/add_clip_image_search.sql
@@ -71,7 +71,10 @@ RETURNS TABLE(
     image_url       TEXT,
     category        TEXT,
     product_type    TEXT,
-    base_price      NUMERIC,
+    -- Phải khớp CHÍNH XÁC kiểu của products.base_price (integer), nếu không
+    -- Postgres báo: "Returned type integer does not match expected type numeric".
+    -- Đã kiểm tra bằng PostgREST OpenAPI, không đoán.
+    base_price      INTEGER,
     similarity      FLOAT
 )
 LANGUAGE plpgsql
@@ -86,7 +89,7 @@ BEGIN
         ce.image_url,
         p.category::TEXT,
         p.product_type::TEXT,
-        p.base_price,
+        p.base_price::INTEGER,
         (1 - (ce.embedding <=> query_embedding))::FLOAT AS similarity
     FROM public.cake_embeddings ce
     JOIN public.products p ON p.id = ce.product_id
